Update slideshow keybindings
The old binding to increase the slideshow delay `.` overrode the dot
command to repeat the last command. For consistency the binding to
decrease the slideshow delay `,` and to start the slideshow `s` were
also changed. The different slideshow bindings now all start with `s`
and require an additional keypress.

fixes #35