Remove paste commands
This is a simple fix for the clash with the binding to the previous
search match `P`. As both `p` for previous image and `P` for previous
search match are used there seems to be no intuitive binding for the
paste commands left.

Adding a different binding at some point is still an option.

fixes #43