Add keyhint widget that displays partial matches
The widget is triggered when a keybinding is partially completed, e.g.
'g' for 'gg' and shown after a configurable delay. The content is the
list of possible completions for the partial, e.g. 'gg', 'gl', ... with
the name of the corresponding command.

The idea behind the widget is to make complicated keybindings more
discoverable.