Update plugin setting from configuration file
The plugins are now stored and set using a dictionary storing key in the
configuration file as plugin name and value in the configuration file as
additional information. The additional information may become useful in
the future, e.g. for plugin managers as URL to the plugin. Plugins can
be deactivated by commenting them which is not more difficult than
switching the value to false.