Reset exif orientation tag when saving images
The orientation is applied currectly when loading using QImageReader.
When saving the orientation was either auto-applied or set by the user
and should therefore be reset to 1 (normal).

To clean up the exif based code a new module imutils.exif was
introduced. All exif related functionality should now be implemented in
this module. This allows putting the piexif import there and only having
to deal with the additional dependency in one module. If the exif module
used as backend is ever changed, it must only be changed in one file.