Add version module to get additional version info
This makes the `vimiv -v` command more useful as it now prints
additional information such as the python, Qt and PyQt version as well
as whether support for optional modules is enabled. It may also be used
for a `:version` command later.

The unit tests for the optional support required also adding pytest to
the test environment in tox.