Make version pop up non-blocking
Instead of running exec_ it is now associated with the mainwindow and
started with show. This allows using vimiv as normal and keeping the
version window open at the same time. In addition testing becomes easier
to handle as the version pop up no longer stops the test chain from
continuing.