Remove strconvert module
It was only used in the settings and it makes more sense for the setting
itself to know how to convert a value to one it can use. This now also
makes it possible to set the setting value to the correct type without
having to go the detour of making it a string first.