Add basic mark class
The class allows marking and unmarking images. When an image is (un-)
marked a corresponding signal is emitted to allow other parts to connect
to this signal and react accordingly, e.g. by highlighting a path in the
library. In addition the `:mark-clear` and `:mark-restore` commands and
two status modules are implemented.