Refactor api.modes moving more logic into class
Similar to the changes done to the settings in cda0212 this removes the
need of many if clauses all over the place as the signals are now in the
mode itself. In addition the functions always required the associated
mode object as argument, it is therefore reasonable to make it a method
of the object instead.