Use custom whitespace escaping in completion
Instead of shlex.quote whitespace is escaped using the custom functions
implemented in utils. Instead of
path with space -> 'path with space'
we now get
path with space -> path\ with\ space
which allows continuing with a regular slash afterwards and keeping
consistent completion.

In addition the ~ character is no longer quoted and therefore fixes
completion in the form of ~/...

To allow proper filtering with the backslash character, the QRegExp now
uses QRegExp.WildcardUnix.

fixes #72