Refactor manipulate to remove some stringly typing
Both the gui manipulate class and the immanipulate Manipulator used the
same names and communicated with each other using names (e.g.
"brightness"). This is not quite safe and a weird form of coupling.

The gui widget now only creates a layout and the display frame for all
manipulations, everything else is dealt with in an according Manipulate
class and in the Manipulator itself.