Update manipulate to allow multiple changes
This refactoring makes it possible to edit the individual tabs and
applying the combined changes to the final image. The manipulations are
now grouped into their own class which associates them to the
corresponding C-function. This ensures correct coupling without the
Manipulations class having to know how every single group works. All
changes are stored in the Manipulator and finally applied in series to
the full-sized pixmap.