Restructure creation of gui.mainwindow
All widgets are now imported and created in the constructor of the
mainwindow. Widgets that are required by children, e.g. the statusbar
required by the complete bar, are passed to them via their constructor.

This centralizes the creation of widgets in one place and keeps the
individual widget modules much more decoupled as they come together in
one single place.