Fix appending to setting value completions
Due to the completion model never being cleared, when changing a
setting, the updated completions were simply added to the existing ones.
When changing a setting multiple times, this lead to a complete mess of
many many suggestions.

Fixed by clearing the previous completions before adding the relevant
current ones.