Fix chaining for arbitrarily nested aliases
The alias is now only applied once all the splitting has been completed.
This ensures that aliases which match others in part are not replaced
preliminary.

Consider, e.g.  ``:q && quack`` where
``q`` is an alias for ``quit`` and
``quack`` is an alias for ``nop``.

Before this would have been expanded to ``:quit && quituack`` in a first
step destroying the second alias. This is now fixed by recursively
splitting and expanding all the aliases until there is no more separator
and alias left.