Implement general and module-dependent logging
It is now possible for each module to have its own logger. How this is
done is shown examplary in the startup module.

This allows filtering and debugging on a module basis and therefore also
allows adding much more debug log statements without making parsing the
content impossible.

All debug log statements will be ported to this interface.