Consistently only read one configuration file
When passing `--config` now only the passed configuration is read.
Before the default user configuration file was read beforehand and the
passed config path only overrode the options there. This was
inconsistent with the behaviour of `--keyfile` and seems to make no
sense as the user explicitly wants a different file.