Slightly refactor completion model handling
The `on_enter` method is now always called when a new completion model
is entered, not when the command line is entered. This allows lazy
initialization of expensive models (e.g. external commands) and
reloading the models more consistently on demand (e.g. trash
completion).

When command mode is left, the completion model is reset to Empty to
ensure on_enter is always called when entering command mode.