Add custom configparser for external resourses
The parser is set-up in a way that it updates values with external
resources. Variables are defined by the syntax `${PREFIX:variable}`.
Currently the only option is `${env:variable}` which is replaced by the
value of the environment variable $env. Extending the options is simple
and can be done in future versions.