Further rewrite starting of end2end tests
This continues the improvements started in
c001eea362a3823830e43818e9bbe6d153d1cbff.

Now the actuall startup is done in the corresponding bdd given steps,
while the cleanup and patching is performed by pytest fixtures. This
removes the need of a vimivprocess module alltogether and the use of a
context manager.