Update alignment of commandline and completion
Before, the completion widget was slightly too large due to padding.
This lead to the possibility of scrolling in horizontal direction and
therefore to misalignment. This is fixed by removing left and right
padding from completion, statusbar and command line. Horizontal padding
is emulated by prepending a space to all completion options. This also
ensure alignment after the ":" character in the command line.