Add new imageformats plugin disabled by default
This plugin makese it easy to add new image formats without impacting
the performance of `utils.files.is_image` by adding very many formats by
default.

The plugin can be activated by adding `imageformats = name, ...` to the
plugins section of the configuration file. As a proof of concept, the
"cr2" format is added here.

To extend support for additional formats, we now only have to:

* Implement a imghdr like test function in imageformats.
* Add that function with the corresponding format name to the FORMATS
  dictionary.
* Activate it in the configuration file as described above.

This is related to supporting raw images (#16) as this is in principle
possible with qt raw, but we don't recognize the images in the correct
format as they are not part of imghdr.