Tests: Add a helper counter for end2end testing
The counter fixture yields a clean instance of the Counter class. Sole
purpose is to provide a command whose result, namely increasing the
number, is easily testable without depending on other vimiv features or
commands. The number is stored as class attribute to avoid having to
deal with class instances in the object registry.