Rename and refactor completion BaseFilter
The filter is now always used as the proxy model. As inheriting is no
longer done, the name should also reflect this. To remove the need of
initializing the same proxy model multiple times, it is now stored in
the completion widget and only the source model containing the data is
changed.