Replace future-task with QTimer-throttle
The new implementation uses a QTimer internally to only call the last
function call after a timeout. The advantage compared to the
future-based implementation is much better readability, increased safety
as we do not keep track of all future timeouts and saving the import of
concurrent.futures. Disadvantage is that we are no longer able to return
values from throttled functions. As this was not used anywhere, this
does not seem like a major downside.