Add test for segfault when manipulating an image
Seems like the segfault only occured when the image size was larger than
the screen size. This was never the case in our end2end tests as we use
small images for fast test speed.

The reason behind this behaviour is still unclear to me. It happens deep
in the paintEvent of `gui.image` in the C++ code of Qt. Not sure if
further investigations would bring more benefit.

Fixes #177