Move pixmap handling into dedicated module
The simple `Pixmaps` storage class was extended with some of the logic
that used to be in the `ImageFileHandler`. Mainly the parts of
retrieving and setting the different pixmaps related to the current
image. This allows not passing the handler, but the pixmaps class, to
manipulate and transform. Therefore the weak reference and cyclic
referencing for type annotating is removed.