Refactor hierarchy of gui.manipulate
The Manipulate class creates the ManipulateImage and handles its
geometry instead of the main window. This removes the need of passing
Manipulate to ManipulateImage and both classes updating their geometry
independently. As they are very much coupled, this is now done in a
single call by Manipulate. To avoid unnecessary calls, the geometry is
only updated when the Manipulate class is visible.

It also avoids having to reference the Manipulator via instance in two
independent places. Instead it is now passed to the ManipulateImage upon
construction and only referenced by Manipulate.