Fix opening single hidden image upon startup
The special handling of passing a single file, i.e. opening all images
in that directory, lead to hidden images being filtered out when opening
as they are not in the filelist when `library.show_hidden` is set to
false. This is fixed by explicitly adding the single image to the
filelist in case it is not there.