NEW: read metadata based on py3exiv2
Additionally to `piexif`, allow reading metadata using `py3exiv2`. This
required adding it as an optional requirement of vimiv. If only `piexif`
is installed, the user is notified that this module will be depreciated.
The exif dict of `ExifInformation` contains no longer the exif value as
a string, but a tuple containing the name of the tag as well as the tag
value in human readable format.