IMPROVE: Mark Speed in Thumbnail Mode
When an image is marked in thumbnail mode, the view is
repainted right after each mark. This makes marking of many images (e.g. `:mark
*` in thumbnail view painfully slow.

To improve that, we add an additional signal which is broadcast from
mark methods in `api._mark` once they are done marking all images. This
way clients can prolong the expensive repainting.