Add :metadata-list-keys command
The command is used to display a list of valid metadata keys for the
current image. As it has two optional arguments, one to customize the
number of columns used, one to print to the terminal, it is implemented
as its own command instead of as an argument to :metadata.