Always chdir into realpath, not the symlink
This fixes the same inconsistency that
0cc236327662dbf18c5f1d82cee287941d4e1350
attempted to solve. As it is no longer possible to enter a symlinked
directory, the stored list of paths always has the real path as base
directory, never the symlink.

The approach here has the advantage of keeping symlinks as-is for
images, thus not introducing any bugs or confusion here.