Adjust thumbnail sizes according to content
The width is now the maximum of the icon size or the current text width.
The height is now the sum of the visible content, i.e. icon and text.

In addition, the icon is no longer perfectly centered in case text is
visible in the bottom, precisely to account for the text.