Refactor thumbnail display options
Instead of setting "when the elements are visible" we now set "what is
visible in each mode". This now leaves us with the two settings
`thumbnail.display_icon` and `thumbnail.display_filmstrip`. Each of them
has the valid options `icon`, `name` and `both`.

This has a few advantages compared to the previous implementation:
- It is not possible to show nothing.
- It is easy to toggle through the valid options.

In addition, the logic checking if a specific element is visible in the
current mode has been put into properties to simplify this in each part
of the code.