Drop python3.6 from CI
Python 3.6 has reached its end of life quite a while ago and is thus no
longer properly supported by many of our dependencies. CI is also
currently broken for python3.6 with the "newer" PyQt5 versions.

In the future we need to also add the newer python versions, i.e. 3.9
and 3.10, to our CI pipeline. However, this is not straightforward due
to py3exiv2 and will only be done after #467 is merged.