Add thumbnail.save configuration option.
If this option is set to False, then do not save newly-generated
thumbnails to the disk. Previously saved thumbnail images will be
used normally and will never be deleted.

This saves space and reduces drive writes at the cost of extra
computational overhead by making the original image be reopened each
time the thumbnail is loaded.