Add ranges for thumbnail loading.
In order to reduce memory usage, the new 'thumbnail.max_ahead'
and 'thumbnail.max_behind' can be used to specify a limit to the
number of thumbnails after the current selection that can be loaded.
Thumbnails outside of this range will be unloaded, and will be loaded
again when they enter the range. Setting either of these variables to
0 will remove the loading limit, to restore previous behavior.