Add the 'none' sorting setting
For sort 'image_order' and 'directory_order', the 'none' setting
can be used to allow images to follow the last order they were
encountered in by the software, instead of re-sorting.

Note that setting another sorting mode while the software is running
then switching back to 'none' will see files remain in the previous
sort order. The original order of images is *not* tracked.

Additionally, 'none' sort cannot be reversed.