Drop note on using Protocols for _ModeWidget
Unfortunately using a Protocol is not enough as we need the _ModeWidget
to be both a QWidget and support the corresponding (Protocol) members.
Combining Protocols and regular inheritance is not possible, and there
currently is no option to combine types in an AND form (equivalent to
the OR form using typing.Union).