Fix keybinding of more than 2 chars were not possible
Keybindings of more than 2 chars were not possible, as the
`TempKeyStorage` was cleaned on every access to `get_keys`. `get_keys`
is invoked by every call of `_process_event`. Removing the call to
`clean` in `get_keys` leaves the code correct, as the keystore is
actively cleaned when no binding matches or on a timeout.